// arch/payload-types.ts
export interface Link {
  label: string;
  type?: 'page' | 'url' | 'phone_number' | 'email';
  icon?: "Thumbs Up" | "Plus" | "Chat Bubble" | "Star" | "Phone" | "Person" | "Map Pin" | "Smiley Face" | "Sad Face" | "Arrow Right" | "Arrow Left" | "Arrow Up" | "Arrow Down" | "Chevron Right" | "Chevron Left" | "Chevron Up" | "Chevron Down" | "Building" | "Calendar" | "Credit Card" | "Envelope" | "Document" | "Globe" | "Fire" | "Heart" | "Link";
  page?: string | Page;
  url?: string;
  open_in_new_tab?: boolean;
  phone_number?: string;
  email?: string;
  id?: string;
  blockName?: string;
  blockType: 'link';
}

export interface Hero {
  heading: string;
  subheading?: string;
  ctas: Link[];
  image: Media;
  id?: string;
  blockName?: string;
  blockType: 'hero';
}

export interface Benefits {
  benefits: {
    icon: "Thumbs Up" | "Plus" | "Chat Bubble" | "Star" | "Phone" | "Person" | "Map Pin" | "Smiley Face" | "Sad Face" | "Arrow Right" | "Arrow Left" | "Arrow Up" | "Arrow Down" | "Chevron Right" | "Chevron Left" | "Chevron Up" | "Chevron Down" | "Building" | "Calendar" | "Credit Card" | "Envelope" | "Document" | "Globe" | "Fire" | "Heart" | "Link";
    benefit: string;
    description?: string;
    id?: string;
  }[];
  id?: string;
  blockName?: string;
  blockType: 'benefits';
}

export interface Testimonials {
  testimonials: {
    rating: number;
    quote: string;
    name: string;
    caption: string;
    id?: string;
  }[];
  image: Media;
  id?: string;
  blockName?: string;
  blockType: 'testimonial-slider';
}

export interface TeamSlider {
  preheading?: string;
  heading: string;
  subheading?: string;
  ctas: Link[];
  id?: string;
  blockName?: string;
  blockType: 'team-slider';
}

export interface CTA {
  preheading?: string;
  heading: string;
  subheading?: string;
  ctas: Link[];
  testimonials: {
    rating: number;
    quote: string;
    name: string;
    caption: string;
    id?: string;
  }[];
  image: Media;
  id?: string;
  blockName?: string;
  blockType: 'cta';
}

export interface FAQs {
  preheading?: string;
  heading?: string;
  subheading?: string;
  faqs: {
    question: string;
    answer: {
      [k: string]: unknown;
    }[];
    id?: string;
  }[];
  cta: {
    heading?: string;
    subheading?: string;
    ctas: {
      label: string;
      icon?: "Thumbs Up" | "Plus" | "Chat Bubble" | "Star" | "Phone" | "Person" | "Map Pin" | "Smiley Face" | "Sad Face" | "Arrow Right" | "Arrow Left" | "Arrow Up" | "Arrow Down" | "Chevron Right" | "Chevron Left" | "Chevron Up" | "Chevron Down" | "Building" | "Calendar" | "Credit Card" | "Envelope" | "Document" | "Globe" | "Fire" | "Heart" | "Link";
      type?: 'page' | 'url' | 'phone_number' | 'email';
      page?: string | Page;
      url?: string;
      open_in_new_tab?: boolean;
      phone_number?: string;
      email?: string;
      id?: string;
      blockName?: string;
      blockType: 'link';
    }[];
  };
  id?: string;
  blockName?: string;
  blockType: 'faqs';
}

export interface PageHeader {
  preheading?: string;
  heading: string;
  subheading?: string;
  ctas: {
    label: string;
    icon?: "Thumbs Up" | "Plus" | "Chat Bubble" | "Star" | "Phone" | "Person" | "Map Pin" | "Smiley Face" | "Sad Face" | "Arrow Right" | "Arrow Left" | "Arrow Up" | "Arrow Down" | "Chevron Right" | "Chevron Left" | "Chevron Up" | "Chevron Down" | "Building" | "Calendar" | "Credit Card" | "Envelope" | "Document" | "Globe" | "Fire" | "Heart" | "Link";
    type?: 'page' | 'url' | 'phone_number' | 'email';
    page?: string | Page;
    url?: string;
    open_in_new_tab?: boolean;
    phone_number?: string;
    email?: string;
    id?: string;
    blockName?: string;
    blockType: 'link';
  }[];
  id?: string;
  blockName?: string;
  blockType: 'page-header';
}

export interface RichText {
  content: {
    [k: string]: unknown;
  }[];
  id?: string;
  blockName?: string;
  blockType: 'rich-text';
}

export interface Testimonial {
  rating: number;
  quote: string;
  name: string;
  caption: string;
  image: Media;
  id?: string;
  blockName?: string;
  blockType: 'testimonial';
}

export interface Locations {
  preheading?: string;
  heading: string;
  subheading?: string;
  locations: {
    name: string;
    opening_hours: {
      sunday: {
        closed?: boolean;
        open?: string;
        close?: string;
      };
      monday: {
        closed?: boolean;
        open?: string;
        close?: string;
      };
      tuesday: {
        closed?: boolean;
        open?: string;
        close?: string;
      };
      wednesday: {
        closed?: boolean;
        open?: string;
        close?: string;
      };
      thursday: {
        closed?: boolean;
        open?: string;
        close?: string;
      };
      friday: {
        closed?: boolean;
        open?: string;
        close?: string;
      };
      saturday: {
        closed?: boolean;
        open?: string;
        close?: string;
      };
    };
    address: {
      line1: string;
      line2?: string;
      line3?: string;
      city: string;
      state: string;
      postal_code: string;
    };
    id?: string;
  }[];
  id?: string;
  blockName?: string;
  blockType: 'locations';
}

export interface TeamGrid {
  id?: string;
  blockName?: string;
  blockType: 'team-grid';
}

export interface OpenPositions {
  preheading?: string;
  heading: string;
  subheading?: string;
  ctas: {
    label: string;
    icon?: "Thumbs Up" | "Plus" | "Chat Bubble" | "Star" | "Phone" | "Person" | "Map Pin" | "Smiley Face" | "Sad Face" | "Arrow Right" | "Arrow Left" | "Arrow Up" | "Arrow Down" | "Chevron Right" | "Chevron Left" | "Chevron Up" | "Chevron Down" | "Building" | "Calendar" | "Credit Card" | "Envelope" | "Document" | "Globe" | "Fire" | "Heart" | "Link";
    type?: 'page' | 'url' | 'phone_number' | 'email';
    page?: string | Page;
    url?: string;
    open_in_new_tab?: boolean;
    phone_number?: string;
    email?: string;
    id?: string;
    blockName?: string;
    blockType: 'link';
  }[];
  image: Media;
  positions: {
    title: string;
    description: string;
    type: 'full-time' | 'part-time' | 'contract' | 'internship';
    location: string;
    close_date?: string;
    id?: string;
  }[];
  id?: string;
  blockName?: string;
  blockType: 'open-positions';
};

export interface PlanPricing {
  preheading?: string;
  heading: string;
  subheading?: string;
  benefits: {
    icon: "Thumbs Up" | "Plus" | "Chat Bubble" | "Star" | "Phone" | "Person" | "Map Pin" | "Smiley Face" | "Sad Face" | "Arrow Right" | "Arrow Left" | "Arrow Up" | "Arrow Down" | "Chevron Right" | "Chevron Left" | "Chevron Up" | "Chevron Down" | "Building" | "Calendar" | "Credit Card" | "Envelope" | "Document" | "Globe" | "Fire" | "Heart" | "Link";
    benefit: string;
    description?: string;
    id?: string;
  }[];
  plan: {
    title: string;
    price: string;
    description: string;
    features: {
      feature: string;
      id?: string;
    }[];
    ctas: Link[];
  };
  id?: string;
  blockName?: string;
  blockType: 'plan-pricing';
}

export interface PricingTable {
  preheading?: string;
  heading: string;
  subheading?: string;
  treatments: Treatment[];
  id?: string;
  blockName?: string;
  blockType: 'pricing-table';
}

export interface ContactOptions {
  options: {
    icon: "Thumbs Up" | "Plus" | "Chat Bubble" | "Star" | "Phone" | "Person" | "Map Pin" | "Smiley Face" | "Sad Face" | "Arrow Right" | "Arrow Left" | "Arrow Up" | "Arrow Down" | "Chevron Right" | "Chevron Left" | "Chevron Up" | "Chevron Down" | "Building" | "Calendar" | "Credit Card" | "Envelope" | "Document" | "Globe" | "Fire" | "Heart" | "Link";
    heading: string;
    subheading?: string;
    cta: Link[];
    id?: string;
  }[];
  id?: string;
  blockName?: string;
  blockType: 'contact-options';
}

export interface ContactForm {
  preheading?: string;
  heading: string;
  subheading?: string;
  image: Media;
  id?: string;
  blockName?: string;
  blockType: 'contact-form';
}
// ...
export interface FeaturedTreatments {
  preheading?: string;
  heading: string;
  subheading?: string;
  ctas: {
    label: string;
    icon?:
      | 'Thumbs Up'
      | 'Plus'
      | 'Star'
      | 'Chat Bubble'
      | 'Phone'
      | 'Person'
      | 'Map Pin'
      | 'Smiley Face'
      | 'Sad Face'
      | 'Arrow Right'
      | 'Arrow Left'
      | 'Arrow Up'
      | 'Arrow Down'
      | 'Chevron Right'
      | 'Chevron Left'
      | 'Chevron Up'
      | 'Chevron Down'
      | 'Building'
      | 'Calendar'
      | 'Credit Card'
      | 'Envelope'
      | 'Document'
      | 'Globe'
      | 'Fire'
      | 'Heart'
      | 'Link';
    type?: 'page' | 'url' | 'phone_number' | 'email';
    page?: string | Page;
    url?: string;
    open_in_new_tab?: boolean;
    phone_number?: string;
    email?: string;
    id?: string;
    blockName?: string;
    blockType: 'link';
  }[];
  featured_treatments: {
    treatment: Treatment;
    image: Media;
    id?: string;
  }[];
  id?: string;
  blockName?: string;
  blockType: 'featured-treatments';
};

export interface BlogPosts {
  id?: string;
  blockName?: string;
  blockType: 'blog-posts';
}

export interface Page {
  id: string;
  title: string;
  slug: string;
  layout: (
    | Hero
    | Benefits
    | Testimonials
    | TeamSlider
    | CTA
    | FAQs
    | PageHeader
    | RichText
    | Testimonial
    | Locations
    | TeamGrid
    | OpenPositions
    | PlanPricing
    | PricingTable
    | ContactOptions
    | ContactForm
    | LatestBlogPosts
    | BlogPosts
    | FeaturedTreatments
  )[];
  meta: {
    title?: string;
    description?: string;
  };
  createdAt: string;
  updatedAt: string;
}

export interface Team {
  id: string;
  name: string;
  slug?: string;
  image: Media;
  role: string;
  short_bio?: string;
  long_bio?: {
    [k: string]: unknown;
  }[];
  index?: number;
  twitter_url?: string;
  linkedin_url?: string;
  website_url?: string;
  createdAt: string;
  updatedAt: string;
}

export interface Treatment {
  id: string;
  title: string;
  description: string;
  slug?: string;
  price?: string;
  category: 'preventative' | 'restorative' | 'cosmetic' | 'orthodontics';
  layout: (
    | Hero
    | Benefits
    | Testimonials
    | TeamSlider
    | CTA
    | FAQs
    | PageHeader
    | RichText
    | Testimonial
    | Locations
    | TeamGrid
    | OpenPositions
    | PlanPricing
    | PricingTable
    | ContactOptions
    | ContactForm
    | LatestBlogPosts
    | BlogPosts
  )[];
  meta: {
    title?: string;
    description?: string;
  };
  createdAt: string;
  updatedAt: string;
}
// ...
export interface Media {
  id: string;
  alt: string;
  url?: string;
  filename?: string;
  mimeType?: string;
  filesize?: number;
  width?: number;
  height?: number;
  sizes: {
    small_landscape: {
      url?: string;
      width?: number;
      height?: number;
      mimeType?: string;
      filesize?: number;
      filename?: string;
    };
    medium_landscape: {
      url?: string;
      width?: number;
      height?: number;
      mimeType?: string;
      filesize?: number;
      filename?: string;
    };
    large_landscape: {
      url?: string;
      width?: number;
      height?: number;
      mimeType?: string;
      filesize?: number;
      filename?: string;
    };
    small_square: {
      url?: string;
      width?: number;
      height?: number;
      mimeType?: string;
      filesize?: number;
      filename?: string;
    };
    medium_square: {
      url?: string;
      width?: number;
      height?: number;
      mimeType?: string;
      filesize?: number;
      filename?: string;
    };
    large_square: {
      url?: string;
      width?: number;
      height?: number;
      mimeType?: string;
      filesize?: number;
      filename?: string;
    };
    small_portrait: {
      url?: string;
      width?: number;
      height?: number;
      mimeType?: string;
      filesize?: number;
      filename?: string;
    };
    medium_portrait: {
      url?: string;
      width?: number;
      height?: number;
      mimeType?: string;
      filesize?: number;
      filename?: string;
    };
    large_portrait: {
      url?: string;
      width?: number;
      height?: number;
      mimeType?: string;
      filesize?: number;
      filename?: string;
    };
  };
  createdAt: string;
  updatedAt: string;
}